Isolate lwjgl timer specifics
